@@ -1,8 +1,6 @@
-'use client';
-
 import { createContext, useContext, useEffect, useState, ReactNode } from 'react';
 import { User, Session } from '@supabase/supabase-js';
-import { supabase } from '../lib/supabase-client';
+import { supabase } from '../lib/supabase';
 import { Profile } from '../types';
 
 interface AuthContextType {
