@@ -1,12 +1,10 @@
-'use client';
-
-
 import { useEffect, useState } from 'react';
-import Link from 'next/link';
+import { Link } from 'react-router-dom';
 import { Trophy, TrendingUp, Coins, Target, ArrowRight } from 'lucide-react';
-import { useAuth } from '@/contexts/AuthProvider';
-import { supabase } from '@/lib/supabase-client';
-import { Match, Token } from '@/types';
+import { useAuth } from '../contexts/AuthContext';
+import { supabase } from '../lib/supabase';
+import { Match, Token } from '../types';
+import DashboardLayout from '../components/DashboardLayout';
 
 export default function DashboardHome() {
   const { profile } = useAuth();
@@ -77,7 +75,8 @@ export default function DashboardHome() {
   ];
 
   return (
-    <div className="space-y-6">
+    <DashboardLayout>
+      <div className="space-y-6">
         <div>
           <h1 className="text-3xl font-bold text-gray-900 dark:text-white mb-2">
             Welcome back, {profile?.full_name?.split(' ')[0]}!
@@ -112,7 +111,7 @@ export default function DashboardHome() {
             <div className="flex items-center justify-between mb-6">
               <h2 className="text-xl font-bold text-gray-900 dark:text-white">Upcoming Matches</h2>
               <Link
-                href="/dashboard/predictions"
+                to="/dashboard/predictions"
                 className="text-green-600 dark:text-green-400 hover:text-green-700 dark:hover:text-green-300 text-sm font-semibold flex items-center"
               >
                 View all
@@ -131,7 +130,7 @@ export default function DashboardHome() {
                 {upcomingMatches.map((match) => (
                   <Link
                     key={match.id}
-                    href="/dashboard/predictions"
+                    to="/dashboard/predictions"
                     className="block bg-gray-50 dark:bg-gray-700/50 rounded-xl p-4 hover:bg-gray-100 dark:hover:bg-gray-700 transition-colors"
                   >
                     <div className="flex items-center justify-between mb-2">
@@ -175,7 +174,7 @@ export default function DashboardHome() {
               <p className="text-sm text-white/75">Perfect for casual predictors</p>
             </div>
             <Link
-              href="/dashboard/tokens"
+              to="/dashboard/tokens"
               className="block w-full bg-white text-green-600 font-bold py-3 rounded-xl text-center hover:bg-gray-100 transition-colors"
             >
               Purchase Tokens
@@ -187,7 +186,7 @@ export default function DashboardHome() {
           <h2 className="text-xl font-bold text-gray-900 dark:text-white mb-4">Quick Actions</h2>
           <div className="grid md:grid-cols-3 gap-4">
             <Link
-              href="/dashboard/predictions"
+              to="/dashboard/predictions"
               className="bg-gradient-to-br from-green-500 to-green-600 text-white rounded-xl p-4 hover:from-green-600 hover:to-green-700 transition-all"
             >
               <TrendingUp className="w-8 h-8 mb-2" />
@@ -195,7 +194,7 @@ export default function DashboardHome() {
               <p className="text-sm text-white/80">Predict upcoming matches</p>
             </Link>
             <Link
-              href="/dashboard/history"
+              to="/dashboard/history"
               className="bg-gradient-to-br from-blue-500 to-blue-600 text-white rounded-xl p-4 hover:from-blue-600 hover:to-blue-700 transition-all"
             >
               <Trophy className="w-8 h-8 mb-2" />
@@ -203,7 +202,7 @@ export default function DashboardHome() {
               <p className="text-sm text-white/80">See your past predictions</p>
             </Link>
             <Link
-              href="/dashboard/support"
+              to="/dashboard/support"
               className="bg-gradient-to-br from-purple-500 to-purple-600 text-white rounded-xl p-4 hover:from-purple-600 hover:to-purple-700 transition-all"
             >
               <Target className="w-8 h-8 mb-2" />
@@ -211,7 +210,8 @@ export default function DashboardHome() {
               <p className="text-sm text-white/80">Contact our team</p>
             </Link>
           </div>
+        </div>
       </div>
-    </div>
+    </DashboardLayout>
   );
 }
