@@ -1,9 +1,10 @@
 import { Calendar } from 'lucide-react';
-
+import DashboardLayout from '../components/DashboardLayout';
 
 export default function EventsPage() {
   return (
-    <div className="space-y-6">
+    <DashboardLayout>
+      <div className="max-w-4xl mx-auto space-y-6">
         <div>
           <h1 className="text-3xl font-bold text-gray-900 dark:text-white mb-2">
             Events & Training
@@ -20,7 +21,8 @@ export default function EventsPage() {
             Event management and training registration will be available soon.
             Check back later for upcoming events!
           </p>
+        </div>
       </div>
-    </div>
+    </DashboardLayout>
   );
 }
