@@ -1,11 +1,9 @@
-'use client';
-
-
 import { useEffect, useState } from 'react';
 import { Trophy, TrendingUp, X, Check } from 'lucide-react';
-import { useAuth } from '@/contexts/AuthProvider';
-import { supabase } from '@/lib/supabase-client';
-import { Prediction } from '@/types';
+import { useAuth } from '../contexts/AuthContext';
+import { supabase } from '../lib/supabase';
+import { Prediction } from '../types';
+import DashboardLayout from '../components/DashboardLayout';
 
 export default function HistoryPage() {
   const { profile } = useAuth();
@@ -70,7 +68,8 @@ export default function HistoryPage() {
   };
 
   return (
-    <div className="space-y-6">
+    <DashboardLayout>
+      <div className="max-w-4xl mx-auto space-y-6">
         <div>
           <h1 className="text-3xl font-bold text-gray-900 dark:text-white mb-2">
             Prediction History
@@ -157,6 +156,7 @@ export default function HistoryPage() {
             </div>
           )}
         </div>
-    </div>
+      </div>
+    </DashboardLayout>
   );
 }
