@@ -1,11 +1,7 @@
-'use client';
-
-
 import { useState } from 'react';
-import Link from 'next/link';
-import { useRouter } from 'next/navigation';
+import { Link, useNavigate } from 'react-router-dom';
 import { Trophy, Mail, Lock, AlertCircle } from 'lucide-react';
-import { useAuth } from '@/contexts/AuthProvider';
+import { useAuth } from '../contexts/AuthContext';
 
 export default function LoginPage() {
   const [email, setEmail] = useState('');
@@ -13,7 +9,7 @@ export default function LoginPage() {
   const [error, setError] = useState('');
   const [loading, setLoading] = useState(false);
   const { signIn } = useAuth();
-  const router = useRouter();
+  const navigate = useNavigate();
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -22,7 +18,7 @@ export default function LoginPage() {
 
     try {
       await signIn(email, password);
-      router.push('/dashboard');
+      navigate('/dashboard');
     } catch (err: any) {
       setError(err.message || 'Failed to sign in');
     } finally {
@@ -34,7 +30,7 @@ export default function LoginPage() {
     <div className="min-h-screen bg-gradient-to-br from-gray-50 to-gray-100 flex items-center justify-center p-4">
       <div className="w-full max-w-md">
         <div className="text-center mb-8">
-          <Link href="/" className="inline-flex items-center space-x-2 mb-4">
+          <Link to="/" className="inline-flex items-center space-x-2 mb-4">
             <div className="bg-beteasy-lime p-3 rounded-2xl">
               <Trophy className="w-8 h-8 text-white" />
             </div>
@@ -99,7 +95,7 @@ export default function LoginPage() {
           <div className="mt-6 text-center">
             <p className="text-gray-600 text-sm">
               Don't have an account?{' '}
-              <Link href="/register" className="text-beteasy-lime hover:text-beteasy-lime-dark font-semibold">
+              <Link to="/register" className="text-beteasy-lime hover:text-beteasy-lime-dark font-semibold">
                 Register here
               </Link>
             </p>
@@ -107,7 +103,7 @@ export default function LoginPage() {
         </div>
 
         <div className="text-center mt-6">
-          <Link href="/" className="text-gray-600 hover:text-gray-900 text-sm transition-colors">
+          <Link to="/" className="text-gray-600 hover:text-gray-900 text-sm transition-colors">
             ← Back to home
           </Link>
         </div>
